+++ core/class-kirki-modules.php
@@ -60,7 +60,6 @@ class Kirki_Modules {
 			array(
 				'css'                => 'Kirki_Modules_CSS',
 				'css-vars'           => 'Kirki_Modules_CSS_Vars',
-				'customizer-styling' => 'Kirki_Modules_Customizer_Styling',
 				'icons'              => 'Kirki_Modules_Icons',
 				'loading'            => 'Kirki_Modules_Loading',
 				'tooltips'           => 'Kirki_Modules_Tooltips',