+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Removed
 * Removed the `branding` module.
+* Removed the `styling` module.
 
 ## 3.0.45 - 2019-09-01