+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 * Completely refactored the google-fonts downloader. Fonts now get stored in `wp-content/fonts/{font-family}/{file}`.
+* Updated Google-Fonts lists.
 
 ### Removed
 * Removed the `branding` module.